# proto/proto.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/device-monitor/{UID}": {
      "get": {
        "operationId": "QueryDeviceMonitor",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/supervisory_controller_serviceQueryDeviceMonitorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "UID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "SControllerService"
        ]
      }
    },
    "/v1/device/{StartTime}/{EndTime}": {
      "get": {
        "operationId": "CensusDevice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/supervisory_controller_serviceCensusDeviceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "StartTime",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "EndTime",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "SControllerService"
        ]
      }
    },
    "/v1/device/{UID}": {
      "get": {
        "operationId": "QueryDevice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/supervisory_controller_serviceQueryDeviceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "UID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "SControllerService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "supervisory_controller_serviceCensusDeviceResponse": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/supervisory_controller_serviceSCResult"
        },
        "TotalDevice": {
          "type": "integer",
          "format": "int32"
        },
        "Devices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/supervisory_controller_serviceSCDeviceMeta"
          }
        }
      }
    },
    "supervisory_controller_serviceQueryDeviceMonitorResponse": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/supervisory_controller_serviceSCResult"
        },
        "Device": {
          "$ref": "#/definitions/supervisory_controller_serviceSCDeviceMeta"
        },
        "Login": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "supervisory_controller_serviceQueryDeviceResponse": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/supervisory_controller_serviceSCResult"
        },
        "Device": {
          "$ref": "#/definitions/supervisory_controller_serviceSCDeviceMeta"
        }
      }
    },
    "supervisory_controller_serviceSCDeviceMeta": {
      "type": "object",
      "properties": {
        "UID": {
          "type": "string",
          "format": "int64"
        },
        "Status": {
          "$ref": "#/definitions/supervisory_controller_serviceSCDeviceStatus"
        },
        "Type": {
          "$ref": "#/definitions/supervisory_controller_serviceSCDeviceType"
        },
        "Meta": {
          "type": "string"
        },
        "CreateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "supervisory_controller_serviceSCDeviceStatus": {
      "type": "string",
      "enum": [
        "SC_STATUS_UNKNOWN",
        "SC_STATUS_ENABLE",
        "SC_STATUS_DISABLE",
        "SC_STATUS_RESTART"
      ],
      "default": "SC_STATUS_UNKNOWN"
    },
    "supervisory_controller_serviceSCDeviceType": {
      "type": "string",
      "enum": [
        "SC_Type_UNKNOWN",
        "SC_Type_LOCK"
      ],
      "default": "SC_Type_UNKNOWN"
    },
    "supervisory_controller_serviceSCErrorCode": {
      "type": "string",
      "enum": [
        "SC_SUCCESS",
        "SC_FAIL",
        "SC_INSERT_DB_FAIL",
        "SC_QUERY_DB_FAIL",
        "SC_TRANSFORM_DATA_FAIL"
      ],
      "default": "SC_SUCCESS"
    },
    "supervisory_controller_serviceSCResult": {
      "type": "object",
      "properties": {
        "Msg": {
          "type": "string"
        },
        "Code": {
          "$ref": "#/definitions/supervisory_controller_serviceSCErrorCode"
        }
      }
    }
  }
}